requirements.txt
Flask
google-api-python-client
google-cloud-pubsub
backoff==1.4.3
ratelimit
google-cloud-resource-manager

